+++ bin/META-INF/persistence.xml
@@ -14,8 +14,9 @@
 		<properties>
 			
 <!-- 			<property name="jakarta.persistence.jdbc.url" value="jdbc:hsqldb:\Users\Brian\Desktop\PAP\BaseDatos\;shutdown=true" /> -->
-				<property name="jakarta.persistence.jdbc.url" value="jdbc:hsqldb:\Users\mwied\OneDrive\Documents\DBeaver\;shutdown=true" />
+<!-- 			<property name="jakarta.persistence.jdbc.url" value="jdbc:hsqldb:\Users\mwied\OneDrive\Documents\DBeaver\;shutdown=true" /> -->
 
+			<property name="jakarta.persistence.jdbc.url" value="jdbc:hsqldb:\Users\fedca\Desktop\PAP\BDPAP\;shutdown=true" />
             <property name="jakarta.persistence.jdbc.user" value="root"/>
 
             <property name="jakarta.persistence.jdbc.password" value="root"/>
