+++ src/presentacion/VtAltaActDep.java
@@ -1,0 +1,15 @@
+package presentacion;
+
+import javax.swing.JFrame;
+import javax.swing.JInternalFrame;
+import javax.swing.JMenu;
+import javax.swing.JMenuBar;
+import javax.swing.JMenuItem;
+
+public class VtAltaActDep extends JInternalFrame{
+	public VtAltaActDep() {
+		this.setClosable(true);
+		this.setBounds(758, 76, 249, 214);
+		this.setVisible(false);
+	}	
+}
