+++ src/presentacion/VtInscClaseDep.java
@@ -1,0 +1,12 @@
+package presentacion;
+
+import javax.swing.JFrame;
+import javax.swing.JMenu;
+import javax.swing.JMenuBar;
+import javax.swing.JMenuItem;
+
+public class VtInscClaseDep extends JFrame{
+	public VtInscClaseDep() {
+		
+	}	
+}
